# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.AdherenceBucket:
    properties:
      bucket:
        type: string
      count:
        type: integer
    type: object
...
  domain.ChallengeStats:
    properties:
      distribution:
        items:
          $ref: '#/definitions/domain.AdherenceBucket'
        type: array
      participation_by_day:
        items:
          $ref: '#/definitions/domain.DayParticipation'
        type: array
    type: object
  domain.ChallengeSummary:
    properties:
      avg_adherence:
        type: number
      best_performer:
        $ref: '#/definitions/domain.ParticipantDetail'
      participants:
        items:
          $ref: '#/definitions/domain.ParticipantDetail'
        type: array
      participants_finished:
        type: integer
      total_checkins:
        type: integer
      total_participants:
        type: integer
    type: object
  domain.DayParticipation:
    properties:
      checked_in:
        type: integer
      date:
        type: string
      total_users:
        type: integer
    type: object
  domain.LeaderboardEntry:
    properties:
      adherence_pct:
        type: number
      current_streak:
        type: integer
      done_days:
        type: integer
      max_streak:
        type: integer
      missed_days:
        type: integer
      total_working_days:
        type: integer
      user_id:
        type: string
      user_name:
        type: string
    type: object
  domain.Notification:
    properties:
      body:
        type: string
      created_at:
        type: string
      data:
        type: object
      id:
        type: string
      is_read:
        type: boolean
      title:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  domain.ParticipantDetail:
    properties:
      adherence:
        type: number
      done_days:
        type: integer
      max_streak:
        type: integer
      user_id:
        type: string
      user_name:
        type: string
    type: object
  domain.PersonalStats:
    properties:
      active_challenges:
        type: integer
      avg_adherence_pct:
        type: number
      finished_challenges:
        type: integer
      max_streak:
        type: integer
      total_challenges:
        type: integer
    type: object
...
  domain.User:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      timezone:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для веб-сервиса коллективного трекинга привычек
  title: Habix API
  version: "1.0"
paths:
...
  /challenges/{id}/leaderboard:
    get:
      parameters:
      - description: Challenge ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.LeaderboardEntry'
            type: array
      security:
      - BearerAuth: []
      summary: Get challenge leaderboard
      tags:
      - stats
...
  /challenges/{id}/stats:
    get:
      parameters:
      - description: Challenge ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ChallengeStats'
      security:
      - BearerAuth: []
      summary: Get challenge statistics
      tags:
      - stats
  /challenges/{id}/summary:
    get:
      parameters:
      - description: Challenge ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ChallengeSummary'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get finished challenge summary
      tags:
      - stats
...
  /notifications:
    get:
      parameters:
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Notification'
            type: array
      security:
      - BearerAuth: []
      summary: List notifications
      tags:
      - notifications
  /notifications/{id}/read:
    patch:
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Mark notification as read
      tags:
      - notifications
  /notifications/read-all:
    patch:
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Mark all notifications as read
      tags:
      - notifications
  /notifications/unread-count:
    get:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
      security:
      - BearerAuth: []
      summary: Get unread notification count
      tags:
      - notifications
...
  /users/{id}/profile:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get user public profile
      tags:
      - users
  /users/me:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
    patch:
      parameters:
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - users
...
  /users/me/stats:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.PersonalStats'
      security:
      - BearerAuth: []
      summary: Get personal stats
      tags:
      - stats
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"